Updated Migration - Changed word id to int from ulong

+++ Floofbot/Services/Repository/Models/FilteredWord.cs
@@ -5,7 +5,7 @@ namespace Floofbot.Services.Repository.Models
     public partial class FilteredWord
     {
         [Key]
-        public ulong Id { get; set; }
+        public int Id { get; set; }
         public ulong ServerId { get; set; }
         public string Word { get; set; }
     }
